Refactor analysis
Break Analyzer up into separate classes for each type of analysis
and add a base class.  Break up the tests accordingly.  Make more
use of the Factory class to create analyzers, including a method
that returns a list of all available analyzers.

This lays the groundwork of refactoring updates to move more code
out of neophile.cli and to avoid running make update-deps twice.
