+++ src/neophile/analysis/helm.py
@@ -1,0 +1,112 @@
+"""Analysis of a repository for needed Helm updates."""
+
+from __future__ import annotations
+
+import logging
+from typing import TYPE_CHECKING
+
+from neophile.analysis.base import BaseAnalyzer
+from neophile.inventory.version import SemanticVersion
+from neophile.update.helm import HelmUpdate
+
+if TYPE_CHECKING:
+    from neophile.inventory.helm import HelmInventory
+    from neophile.scanner.helm import HelmScanner
+    from neophile.update.base import Update
+    from typing import List
+
+__all__ = ["HelmAnalyzer"]
+
+
+class HelmAnalyzer(BaseAnalyzer):
+    """Analyze a tree for needed Helm updates.
+
+    Parameters
+    ----------
+    root : `str`
+        Root of the directory tree to analyze.
+    scanner : `neophile.scanner.HelmScanner`
+        Scanner for Helm dependencies.
+    inventory : `neophile.inventory.HelmInventory`
+        Inventory for Helm repositories.
+    allow_expressions : `bool`, optional
+        If set, allow dependencies to be expressed as expressions, and only
+        report a needed update if the latest version is outside the range of
+        the expression.  Defaults to false.
+    """
+
+    def __init__(
+        self,
+        root: str,
+        scanner: HelmScanner,
+        inventory: HelmInventory,
+        *,
+        allow_expressions: bool = False,
+    ) -> None:
+        self._root = root
+        self._allow_expressions = allow_expressions
+        self._scanner = scanner
+        self._inventory = inventory
+
+    async def analyze(self) -> List[Update]:
+        """Analyze a tree and return a list of needed Helm changes.
+
+        Returns
+        -------
+        results : List[`neophile.update.base.Update`]
+            A list of updates.
+        """
+        dependencies = self._scanner.scan()
+        repositories = {d.repository for d in dependencies}
+        latest = {}
+        for repo in repositories:
+            latest[repo] = await self._inventory.inventory(repo)
+
+        results: List[Update] = []
+        for dependency in dependencies:
+            repo = dependency.repository
+            name = dependency.name
+            if name not in latest[repo]:
+                logging.warning(
+                    "Helm chart %s not found in repository %s", name, repo
+                )
+                continue
+            if self._helm_needs_update(dependency.version, latest[repo][name]):
+                update = HelmUpdate(
+                    name=name,
+                    current=dependency.version,
+                    latest=latest[repo][name],
+                    path=dependency.path,
+                )
+                results.append(update)
+
+        return results
+
+    def name(self) -> str:
+        return "helm"
+
+    def _helm_needs_update(self, current: str, latest_str: str) -> bool:
+        """Determine if a Helm dependency needs to be updated.
+
+        Parameters
+        ----------
+        current : `str`
+            The current version number.  If this is not a valid version number,
+            it is assumed to be a match pattern.
+        latest_str : `str`
+            The version number of the latest release.
+
+        Returns
+        -------
+        result : `bool`
+            Whether this dependency should be updated.  Returns true if the
+            current version is invalid or if it is older than the latest
+            version.
+        """
+        latest = SemanticVersion.from_str(latest_str)
+        if SemanticVersion.is_valid(current):
+            return latest.parsed_version > current
+        elif self._allow_expressions:
+            return not latest.parsed_version.match(current)
+        else:
+            return True
